+++ src/main/java/wellink/test/task/entities/materials/ThreadRoll.java
@@ -16,11 +16,11 @@ public class ThreadRoll extends Material {
 
     private Density threadDensity;
 
-    public ThreadRoll(Integer id, String name, Double price, Color color) {
+    public ThreadRoll(Long id, String name, Double price, Color color) {
         super(id, name, price, color);
     }
 
-    public ThreadRoll(Integer id, String name, Double price, Color color, Density threadDensity) {
+    public ThreadRoll(Long id, String name, Double price, Color color, Density threadDensity) {
         super(id, name, price, color);
         this.threadDensity = threadDensity;
     }
