+++ src/main/java/wellink/test/task/repositories/materials/ButtonRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.Button;
 
-public interface ButtonRepository extends CrudRepository <Button, Integer> {
+public interface ButtonRepository extends CrudRepository <Button, Long> {
 }
