+++ src/main/java/wellink/test/task/repositories/materials/LaceRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.Lace;
 
-public interface LaceRepository extends CrudRepository <Lace, Integer> {
+public interface LaceRepository extends CrudRepository <Lace, Long> {
 }
