+++ src/main/java/wellink/test/task/repositories/materials/ShoeSoleRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.ShoeSole;
 
-public interface ShoeSoleRepository extends CrudRepository <ShoeSole, Integer> {
+public interface ShoeSoleRepository extends CrudRepository <ShoeSole, Long> {
 }
