+++ src/main/java/wellink/test/task/repositories/materials/ThreadRollRepository.java
@@ -3,5 +3,5 @@ package wellink.test.task.repositories.materials;
 import org.springframework.data.repository.CrudRepository;
 import wellink.test.task.entities.materials.ThreadRoll;
 
-public interface ThreadRollRepository extends CrudRepository <ThreadRoll, Integer> {
+public interface ThreadRollRepository extends CrudRepository <ThreadRoll, Long> {
 }
